Don't go to the Diorama Table when typing GO TO BUREAU.
Add GIFTSHOP as a synonym for the Gift Shop.

Add some regions and rules.

Add room-restriction rule for Palm Square and beyond.

Make the unmonkey test command liberate movement outside the starting area as well.

Now that we restrict every destination outside the starting area, we must make the restrictions easier to remove as well.

Check whether the plans are seen rather than handled.

This should fix the issue when players letter-transform the plans before touching them.

Add room-restriction rule for rooms 'before' cute security door.

Give a response to trying to approach any parts of the game outside the cute security door, so that the player won't be transported to the door and then stopped.