Fix an issue where a room description would be printed twice.
There are some places where the game halts you if you try to approach a location beyond it, because something important happens. But if the room where you are halted is in fact your final destination, a the room description would be printed twice. This commit checks if the room is the final destination before halting you.