Allow the player do leave things on a supporter inside the car in the traffic circle.
The Traffic Circle is privately-controlled, i.e. we are not allowed to drop things there so as not to make essential things unavailable. However, we should still be allowed to leave things in the car. There already is code in place to let us leave things in containers inside the car, and this is the corresponding allowance for putting things on supporters in the car.