Make an exception for the synthesizer when creating fluid things.
We usually don't allow creating fluids inside container that already contain other, potentially water-sensitive things, but we make an exception for the synthesizer so as not to inconvenience the player too much. Let's assume that the synthesizer has a wide flat bottom so that the liquid stays in one end, separate from the other stuff.