Prefer peripheral quips when disambiguating.
This is mainly to elminitate the case >TOPICS

We could ask for privacy or ask whether she had trouble with customs.

>ASK ABOUT CUSTOMS

What would you like to discuss: 1) where customs is found or 2) whether she had trouble with customs?

but there are probably similar cases elsewhere.

The Inform 6 code is of course not really all that much better or faster, but now that it is written it would be a pity not to use it.