This should fix a rather serious limitation of Scope Caching.
I should have done this a long time ago. Up until now, only things in the previous location would be marked-visible when first walking to a new location. The reason it hasn't caused more problems is that I have replaced "marked-visible" with  "enclosed by location" in a lot of places.