Homonym using an index rather than the description
This is a better fix for #111. Previously, objects with random
variation in their description might be repeatedly homonym-paddled
into the same kind of object. This commit gives every thing a
"homonym index", which is the same for all instances of a kind.
The homonyming code will check this number and disallow homonyming
between two things with the same homonym index.