Change all input to lower case
Doing this in the most obvious way (let N be "[the player's
command in lower case]"; change the text of the player's command to
N;) caused quite a performance hit, and while investigating this I
learned that changing the text of the player's command will re-run
the tokenisation step, which can be quite slow. In order to make
this happen less frequently, I implemented a new rulebook that runs
at the beginning of the After reading a command stage, copies the
player's command to a global string (player-command-substitute),
operates on this instead of the "player's command" array, and then
changes the text of the player's command just once when all changes
are done.

As it turns out, most of the performance hit could be avoided by
changing 'let N be "[the player's command in lower case]"' to "let N
be the player's command; let N be N in lower case;" ("let N be the
player's command in lower case;" was also a lot faster, but not quite
as much). Also, after extensive profiling I found that only a few
"after reading a command" rules actually became faster by being turned
into "command-string altering" rules, which makes all this complication
and uglification less obviously worth it. Still, it *is* a performance
gain and doesn't break anything.