Clarify that we get arrested when we do
When the player tries to walk past the secretary a second time
without an invitation, they are apprehended and the game is over.
If the player chooses to be "resurrected", the message "Our
detention goes badly" is never printed, so a clarifying sentence
is added. ("She sends the room into lockdown".)