Understand trying to sit, stand or recline on unsuitable things
The bug that this change was meant to fix (commit 063e741, run-time
error when trying to check the posture of the sky) does not seem to
occur anymore. I would also inadvertently give the message
"I can't see any such thing" if the player tried to sit or stand
on something non-enterable.