Fix some more bugs in Postures
If you typed LIE ON CHAIR when the chair does not allow reclining,
you would still enter the chair, despite getting a rejection
message. Also fix some other inconsistencies and cosmetic things.

To make this a proper reusable extension, we still need to add
checks to see if the actor is visible before printing any
messages. We skip that here for speed reasons.