Always understand LIE as a verb in the Dormitory
In order to understand LIE as referring to conversation quips,
the input would be hacked to read LIEABOUT instead. This would
sometimes lead to the message "That is not a verb I understand"
when trying to lie down on the dormitory bed. We fix this by
only hacking the input when the relevant quips are available.