Speed up Numbered Disambiguation Choices
Use a kind of value instead of a number for the disambiguation id.
It might break if there are more than 12 disambiguation choices,
but I'm not sure that can actually happen.

Thanks a lot to Andrew Plotkin for this code.
See https://www.intfiction.org/forum/viewtopic.php?f=7&t=26307