Improved the new Numbered Disambiguation Choices again
This is more robust and readable. Thanks again, Dannii!

It will still break if there are more than 12 choices,
such as when typing PURLOIN MAN in a debug build,
but I think that is acceptable, at least until someone
can show a way to trigger it in a normal game.