Yes-no questions usually have a double more-than sign
Except that the "Please answer yes or no.>" follow up has
a single. It is also somewhat inconsistent whether there
should be a space after the more-than prompt or not.

Note to self: Lectrote will hang if the window is resized while
this question is active (Quixe run: gli_put_char: window has
pending line request.)