Remove line break-suppressing space
This was added in order to suppress an extra paragraph break when
"approaching" Alex's apartment for the first time. But then there
would be a missing break when trying to go there normally, i.e. by
typing SW in Palm Square. So let's try to live with that double
break instead. It is no big dieal.