Adjustments to automatic descriptions of newly created objects
This makes the descriptions of new things created by letter
transformations more consistent. Now things created by the
synthesizer, the T-inserter and the cryptolock will be
described automatically the first time you create them, but only
the first time. Before, there were no description of products of
the cryptolock. Also, the descriptions of the products of the
T-inserter have been moved to a more logical place.
Note that homonyms are still described every time, to avoid
confusion as to what exactly was created.