Make animals stay on the dais for a while
This makes animals and fake people on the dais behave like when
they enter a car with you: eventually they will exit, if a random
chance of 1 in 10 succeeds, but not on the same turn that they
entered. This is an uneasy compromise: sometimes the player is
trying to do something which requires a creature to stay in place,
at other times something which requires the creature to get off.
Not all creatures can be picked up, and it is not always viable
to turn them into something inanimate. We also don't want to
implement creatures following orders from the player. The randomness
means that if you are trying to do something that requires an animal
to stay put for 10 turns, you can in theory keep trying and eventually
it will succeed.