Make disabled line echo and graphics survive restore
If we restore a save file made on an interpreter that does not
support line echo suppression or graphics, we still want to suppress
echo and show graphics if supported. If we switch off the map, we
probably want it to stay off even if we restore a save made with
graphics on or restart the game.
This also fixes reloading achievements after restoring.