Replace mention-lists with i6 array
Instead of every quip having a list property of subjects, we put
all subjects into a single Inform 6 array and gives the quips
pointers to sequences in it that provide the same information.

This makes the code less readable and harder to change, but makes up
for it with improved performance. To make it easier to add new
subjects if needed in the future, some disabled by default code is
provided in the file Regenerate Mentions.i7x to generate a new array
of subjects and new indices for every quip.