Fixes to Flexible Windows to preserve default colours
Before, text colour would always be set to black on white after
restarting, overriding the interpreter settings. On Gargoyle, it
would also set the cursor to black and window padding to white.
We now preserve the main windows on restart and don't
try to counteract the default behaviour of Gargoyle.