Do not disfavor previously-created things. Fixes #129
There was a bug in "the intelligent substitution rule" which made it
disfavor previously created items when there were several identical
to choose between. All in-between stages in a creation chain, such
as the card in CHARD -> CARD -> CAR, retain all their proffering
relations when the final product is gelled back to its original object.
This made the code think they were part of some different creation
chain, and they were disfavored in order to avoid bugs like the
disappearing as.

This should have created a lot more problems than it seems to have
done.