Rework local looking
Local looking is the action that redirects EXAMINE <room> to LOOK.
Previously it would confuse the parser when when trying to examine
multiple things at once, such as typing EXAMINE ALL or EXAMINE OILS.
Now you get a proper error message: "You can't use multiple objects
with that verb" instead of "You can't see any such thing" or
printing the description of the room.
For some reason it is also a little faster.