Open closed containers when inserting things into them. Fixes #132
This was already implemented for held things, but now it works for
all things. Still won't work with PUT ALL IN, though.
This may not be the most logical place for this code,
but it makes sure that actions are attempted in the right order, so
that we don't open the container even if the first noun can't be
taken.
Also changes the code style in this rule from commas to colons.