Don't let creatures try to exit the repository
In some cases, living creatures may be moved to the repository while
still listed in the people_present array, and then try to exit in
the contained people exiting rule. An example is when turning the
attendant's blouse into a louse (see #139). We fix this by blocking
exiting from the repository in the MyContainedPeopleExiting function.