Let things we force-drop on spinner land on the ground
Because there is only room for one thing on the spinner, we can't drop
anything when standing on it. If we involuntarily dropped something because
we have created a heavy object or a liquid in our hands, we would just
keep holding it. So we move it to the ground.