Always mention picking up things when sitting
If sitting on a chair by typing ENTER CHAIR, the game will tell us
that we pick up anything already on it. If we type SIT CHAIR, however,
the game would pick it up without telling. Now we check if the action
is silent, and print a message if it is.