Highlight the things on the diorama table. Fixes #147
This was surprisingly complicated. I don't understand why the
printing the name of activity does not work with the standard list
writer, but works with the Large Game Speedup provided one. I also
don't get why the [are] adaptive verb does not work with the Large
Game Speedup list writer. But this way everything works and everyone
is happy.