Fix teach scoring rule (I think)
I'm not 100 percent sure about the intent of this rule, but it seems
to check whether we got our first point by waving the letter-remover at
something, and then print the first message, otherwise the second. As we
have to use the letter-remover to get out of the tutorial area, this
will always be our first point, and the second message would never be
printed.

However, when we wave the U-remover at the mourning dress,
the current action will be "waiting" when we get to this check, so
the second message was always printed rather than the first. Perhaps
this was different in earlier versions of Inform. The check and one
of the messages is still redundant.