Make things seen when waving remover at them
Now that we don't create a new string by printing the name of the second
noun when letter-removing, the second noun does not automatically get its
seen flag set. This is because the seen mechanisms relies on the printing
the name of activity. This creates some harmless differences that screws
with regression tests, so let's restore the previous way like this.