Pre-remove achievements option before ending the game
Things will break if we blank out a row in the Table of Final Question
Options after the game has ended, but we don't want to award the end-game
achievements before it has ended. So we will have to check which achievements
we are going to award beforehand, and blank out the row if the player will
have all of them.