Add subject rum
This is to make the command A RUM work for performing the quip
"challenge Parker about the rum". This command is used in the
test command TEST BARMAN and works in at least release 6.

I'm not sure how it ever did work, as commands such as
ASK ABOUT [a quip] or A [quip] are supposed to only work
with questioning quips, and "challenge Parker..." is a
performative quip.