Acknowledge secret door before we have opened it
If we tried to go through the secret "odor" door after having created it
but before opening it, we would get the "None of the doors in that direction
look useful to us" response. We fix this by checking the seen property of the
door instead of whether we have visited the Oracle Project.

Reported by @OtherOlly in #157.