Add a message when examining mirrors while wearing the monocle
It does not make a whole lot of sense that examining and looking into
(searching) mirrors while wearing the monocle are redirected to the same
action, while LOOK AT MIRROR THROUGH MONOCLE is a different action with
a different response.

(I guess it would have made more sense if EXAMINE MIRROR and LOOK INTO
MIRROR would have been different actions.)

Also, examining things when wearing a monocle usually tells us whether they
are real or not, with the exception of mirrors.

So to make things more consistent, we append the "looking at a mirror though
the monocle" message to the end of all variants of searching a mirror while
wearing the monocle.