Let the bull trigger the cryptobucket safety override
The description hints that the bull is supposed to be a threat to
the player, so it seems wrong to just have it standing there like a
statue. We could make it an instant death, which would let the player
see the description, but it makes more sense that the cryptolock safety
override prevents us from creating it in the first place.