Add many new regtests
The "tests" regtest suite contains the majority of the internal test
commands converted to regtests. It also tests many of the more
recently discovered bugs.

Only tested with my RemGlk build of Git.

Less robust than the previous regtests, and is likely to break
when things change or when using a different interpreter.

Run using the Bash script regtest-test.sh