Check if written save files are zero length. Fixes #191
By default, file writing fails aren't detected by Inform. This becomes a problem
if we run out of disk space, because the games allows saving without complaint,
and then the save files will be useless when the player tries to restore.

We fix this by rewriting the save game routine to check that the files aren't 0 length.

Fix provided by Dannii Willis.