More Facing adjustments
Replace the can't see through closed door rule in the Facing extension in order
to remove the automatic opening of closed doors when trying to look through them.

Also fix the way adjacent rooms are described from other rooms. We keep the title
case capitalization and say computer-gamey things like "We make out the Precarious
Perch that way."

It would be nice if we could instead use lower case where appropriate, and add the
correct article ("a precarious perch" rather than "the precarious perch" until
the location is visited and so on), but that would be a lot of work and require
new properties to be set for each one of the 100 rooms, so that will have to wait.