Fix articles of rooms
An attempt to make the use of the definite article in room names
consistent. Most of the time, room names are displayed without an
article, but it is used when looking into an adjacent room ("We
make out the roundabout that way"), when listing exits ("We can go
west to Sigil Street from here"), when asking for disambiguation
("Which do you mean, 1) Higgate's office, or the 2) Graduate
Student Office?"), and when reporting fast-travel ("We walk up the
steep switchback of Deep Street and through the perilous traffic
of the roundabout as far as the Monumental Staircase.")

This is currently implemented in a slightly haphazard way, where
the rooms are sometimes explicitly given the "proper-named" or
the "improper-named" property, but in other cases we rely on the
fact that the compiler tries to infer this from whether the room
is referred to with an a definite article the first time it is
mentioned in the source ("The Hostel is east of Heritage
Corner", "West of Long Street North is Fleur d'Or Lobby.")

There is an assumption in several places that if a room has the
proper-named property, it should be capitalized, so "My Apartment"
rather than "my apartment". I considered adding a special flag
for rooms that shouldn't have an article but still be lower-case,
but that stull wouldn't help with room names such as "Slango's
Bunk".