Make room names in sentences look more natural
An attempt to make the use of room names in sentences more natural. This affects
listing exits, "can't go that way" messages, looking into adjacent rooms, and
reporting travelling longer distances.

Changes things like "We can see Slango's Bunk that way" to "We can see Slango's
bunk that way."

Also tries to deal with proper-named rooms that have a definite article, such
as The Counterfeit Monkey or The Aquarium bookstore. Not sure if the approach
used, making them improper-named, is the best one, though.