Don't allow turning Brock into a Bock. Fixes #234
Looks like the bock was added in release 5 as a derivative of brock. We need
to block actually turning Brock into a Bock, though, otherwise the game breaks:
The player can leave the T-inserter early, and later gel the bock. Thanks to
Vivian Thomas (@Roachbones) for the bug report and fix.