Make creating the scree kill the player again
This was changed in 36f0761, mainly in order to make it possible to
see some custom responses to trying to do things to the scree.
However, there are several places in the code where the screen is
assumed to be present in the cinema, and I am not really up
to fixing all of them at the moment.

Thanks to Mike Russo for pointing this out in his Let's play on
intfiction.org.