Update Subcommands extension
This does not actually touch any subcommands functionality, but
because the extension replaces NounDomain function, any changes
to NounDomain must be made here.

What we in fact do is implementing the upstream fix for bug
I7-2115.

I haven't actually seen this bug manifest in Counterfeit Monkey,
so it is a pre-emptive fix. But it is the kind of bug that seems
like it should be possible to trigger in most games.