Fix entering Brock's laptop password verbatim
Punctuation Removal replaces the exclamation mark in the password
with a space, so we check for the result of that instead. (An
alternative would have been to roll our own punctuation removal,
but this is less work.)

This bug was discovered by Mike Russo in the Let’s Play: Counterfeit
Monkey thread on intfiction.org.