Make Brock's nice-garment quip weakly phrased
Apparently there is a small chance that Brock's quip about something
we're wearing ("Nice wig") will be preceded by a beat, and an even
smaller chance that it will be the exact same beat as in the following
paragraph, as Mike Russo discovered in his let's play.

By giving the quip the weakly-phrased property, we make sure that
it will be printed as a part of the following paragraph, hopefully
preventing that this ever happens again. I also think it looks
nicer this way.