Fix switching on a computer while wearing the monocle
Pointed out by Mike Russo in his let's play.

When starting up a computer, we automatically examine an associated
screen. This may give a confusing message if we are wearing the monocle,
such as "The monocle pings happily as you sight Book search with the
crosshairs." (There may be something else going wrong here, as the
Book search is a search engine and not a screen.)

We hackily fix this by temporarily removing the monocle from the player.